+++ src/esmockIsLoader.js
@@ -1,3 +1,0 @@
-import { loaderIsVerified } from './esmockLoader.js'
-
-export default (c => () => (c = c || loaderIsVerified(import.meta.url)))()