+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # changelog
 
+ * 2.3.8 _unreleased_
+   * [reuse moduleid regexp](https://github.com/iambumblehead/esmock/pull/231) replacing separately created regexps
+   * [remove esmockIsLoader.js](https://github.com/iambumblehead/esmock/pull/231) export from esmockLoader.js instead
  * 2.3.7 _Aug.15.2023_
    * [normalize package.json url](https://github.com/iambumblehead/esmock/pull/225) and [reduce loc for loader verification](https://github.com/iambumblehead/esmock/pull/226)
    * [small adjustments](https://github.com/iambumblehead/esmock/pull/228) reducing lines of code